requirements.txt
alabaster==0.7.13
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
billiard==4.2.0
Brotli==1.1.0
celery==5.3.6
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.37.0
colorama==0.4.6
cron-descriptor==1.4.3
Cython==3.0.8
Django==4.2.10
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docutils==0.20.1
drf-spectacular==0.27.0
flower==2.0.1
gunicorn==21.2.0
humanize==4.9.0
idna==3.6
imagesize==1.4.1
inflection==0.5.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kaleido==0.2.1
kombu==5.3.5
MarkupSafe==2.1.3
numpy==1.26.3
packaging==23.2
pandas==2.2.0
patsy==0.5.6
Pillow==10.1.0
plotly==5.18.0
pmdarima==2.0.4
prometheus_client==0.20.0
prompt-toolkit==3.0.43
Pygments==2.17.2
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.1
reportlab==4.1.0
requests==2.31.0
rpds-py==0.16.2
scikit-learn==1.4.0
scipy==1.12.0
sentry-sdk==1.40.6
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
statsmodels==0.14.1
tenacity==8.2.3
threadpoolctl==3.2.0
tornado==6.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
